InstanceListener abstracted out to it's own file as there will be a bit more logic surrounding it with the addition of the union of notifications
